ml_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
tldextract==5.1.1
requests==2.31.0
joblib==1.3.2
